+++ src/protocols/ethereum/components/TransferSendForm.vue
@@ -228,7 +228,7 @@ export default defineComponent({
       return undefined;
     }
 
-    let polling: NodeJS.Timer | null = null;
+    let polling: NodeJS.Timeout | null = null;
 
     onMounted(() => {
       polling = executeAndSetInterval(async () => {